tests/requirements.txt
pytest>=6.2.2
pytest<7.0.0
owlrl>=5.2.1
owlrl<6.0
psycopg2-binary>=2.8.6
psycopg2-binary<3.0
